Rename Android package to net.interstellarai.cosmicmatch + CI signs release APKs
Package rename:
- namespace and applicationId: com.cosmicmatch.cosmic_match → net.interstellarai.cosmicmatch
- MainActivity.kt moved to android/app/src/main/kotlin/net/interstellarai/cosmicmatch/

CI: produce signed release APKs in the Build job
- Every CI run previously shipped an unsigned debug APK (random debug key
  per runner), so each APK had a different signature and reinstalls
  required uninstall first, wiping local data.
- Build job now decodes the release keystore from secrets (same as the
  existing Release job pattern for AABs), writes key.properties, and runs
  flutter build apk --release. All CI APKs share the stable upload key
  signature; installs update in place.
- Artifact renamed: debug-apk → release-apk, retention bumped to 30d.

Verified locally with Flutter 3.41.7 + JDK 17 — flutter build apk --debug
succeeds with the renamed package.
