+++ android/app/build.gradle.kts
@@ -15,7 +15,7 @@ if (keyPropertiesFile.exists()) {
 }
 
 android {
-    namespace = "com.cosmicmatch.cosmic_match"
+    namespace = "net.interstellarai.cosmicmatch"
     compileSdk = flutter.compileSdkVersion
     ndkVersion = flutter.ndkVersion
 
@@ -29,7 +29,7 @@ android {
     }
 
     defaultConfig {
-        applicationId = "com.cosmicmatch.cosmic_match"
+        applicationId = "net.interstellarai.cosmicmatch"
         minSdk = 26
         targetSdk = flutter.targetSdkVersion
         versionCode = flutter.versionCode
